backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autoflake==2.3.1
autopep8==2.3.1
bandit==1.7.9
black==24.8.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
com2ann==0.3.0
cryptography==43.0.0
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.1
django-bootstrap5==24.2
django-filter==24.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django_rest_framework_redocs==0.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docformatter==0.7
drf-yasg==1.21.7
flake8==7.1.1
fonttools==4.53.1
gunicorn==23.0.0
html5lib==1.1
idna==3.7
inflection==0.5.1
isort==5.13.2
libcst==1.4.0
lxml==5.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
pathspec==0.12.1
pbr==6.0.0
pillow==10.4.0
platformdirs==4.2.2
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyflakes==3.2.0
pyformat==1.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pypdf==4.3.1
PyPDF2==3.0.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python3-openid==3.2.0
pytz==2024.1
pyupgrade==3.17.0
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
ruff==0.6.1
setuptools==72.2.0
shed==2024.3.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
ssort==0.13.0
stevedore==5.2.0
svglib==1.5.1
tinycss2==1.3.0
tokenize-rt==6.0.0
typing_extensions==4.12.2
tzlocal==5.2
unify==0.5
untokenize==0.1.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
